@@ -4,39 +4,57 @@
   "private": true,
   "description": "Client for FeelsBox",
   "main": "index.js",
-  "dependencies": {
-    "axios": "^0.17.1",
-    "babel-cli": "^6.26.0",
-    "compression": "^1.7.2",
-    "express": "^4.16.3",
-    "firebase": "^4.8.1",
-    "material-ui": "^1.0.0-beta.40",
-    "material-ui-icons": "^1.0.0-beta.36",
-    "open": "0.0.5",
-    "react": "^16.3.1",
-    "react-color": "^2.14.1",
-    "react-dom": "^16.3.1",
-    "react-router-dom": "^4.2.2",
-    "reactcss": "^1.2.3",
-    "serve-favicon": "^2.5.0"
-  },
   "scripts": {
-    "start": "webpack-dev-server --mode development --open",
+    "start": "webpack-dev-server --mode development --https",
     "build": "webpack --mode production",
     "heroku-postbuild": "npm run build"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.4",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-3": "^6.24.1",
-    "copy-webpack-plugin": "^4.5.1",
+    "@babel/cli": "^7.8.4",
+    "@babel/core": "^7.9.0",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/preset-env": "^7.9.0",
+    "@babel/preset-react": "^7.9.4",
+    "@hot-loader/react-dom": "^16.13.0",
+    "babel-eslint": "^10.1.0",
+    "babel-loader": "^8.1.0",
+    "babel-plugin-react-require": "^3.1.3",
+    "copy-webpack-plugin": "^5.0.1",
+    "eslint": "^6.8.0",
+    "eslint-plugin-class-property": "^1.1.0",
+    "eslint-plugin-react": "^7.19.0",
+    "fs-extra": "^9.0.0",
     "html-webpack-plugin": "^3.2.0",
-    "webpack": "^4.5.0",
-    "webpack-cli": "^2.0.14",
-    "webpack-dev-server": "^3.1.1",
-    "webpack-pwa-manifest": "^3.6.2"
+    "react-hot-loader": "^4.12.20",
+    "webpack": "^4.42.1",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3",
+    "webpack-pwa-manifest": "^4.2.0"
+  },
+  "dependencies": {
+    "@babel/polyfill": "^7.8.7",
+    "@material-ui/core": "^4.9.8",
+    "@material-ui/icons": "^4.9.1",
+    "apollo-cache-inmemory": "^1.6.5",
+    "apollo-client": "^2.6.8",
+    "apollo-link": "^1.2.13",
+    "apollo-link-http": "^1.5.16",
+    "axios": "^0.19.2",
+    "compression": "^1.7.4",
+    "copy-to-clipboard": "^3.3.1",
+    "email-validator": "^2.0.4",
+    "express": "^4.17.1",
+    "graphql": "^15.0.0",
+    "graphql-tag": "^2.10.3",
+    "lodash": "^4.17.15",
+    "open": "^7.0.3",
+    "react": "^16.13.1",
+    "react-apollo": "^3.1.3",
+    "react-color": "^2.18.0",
+    "react-dom": "^16.13.1",
+    "react-router-dom": "^5.2.0",
+    "reactcss": "^1.2.3",
+    "recompose": "^0.30.0",
+    "serve-favicon": "^2.5.0"
   }
 }
