@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {withStyles} from 'material-ui/styles';
-import AppBar from 'material-ui/AppBar';
-import Toolbar from 'material-ui/Toolbar';
-import Typography from 'material-ui/Typography';
+import {withStyles} from '@material-ui/core/styles';
+import AppBar from '@material-ui/core/AppBar';
+import Toolbar from '@material-ui/core/Toolbar';
+import Typography from '@material-ui/core/Typography';
 
 import Navigation from '-/components/Navigation';
 
@@ -23,7 +23,7 @@ function FBAppBar(props) {
             <AppBar position="static">
                 <Toolbar>
                     {includeNav && <Navigation />}
-                    <Typography variant="title" color="inherit" className={classes.flex}>
+                    <Typography variant="h6" color="inherit" className={classes.flex}>
                         {title}
                     </Typography>
                     {iconRenderer}
