@@ -1,35 +1,74 @@
+/* eslint-disable no-undef */
 import React from 'react';
-import {withStyles} from 'material-ui/styles';
-import CssBaseline from 'material-ui/CssBaseline';
-import {CircularProgress} from 'material-ui/Progress';
+import {withStyles} from '@material-ui/core/styles';
+import CssBaseline from '@material-ui/core/CssBaseline';
+import Icon from '@material-ui/core/Icon';
+import Button from '@material-ui/core/Button';
+import Typography from '@material-ui/core/Typography';
+import {Redirect} from 'react-router-dom';
 
 import AppBar from '-/components/AppBar';
 
-const styles = theme => ({
+const styles = () => ({
     root: {
         padding: '10px'
     },
     text: {
         fontSize: 14,
         fontFamily: 'sans-serif'
-    },
-    progress: {
-        margin: '200px auto',
-        display: 'block'
     }
 });
 
 class Login extends React.Component {
+    constructor(props) {
+        super(props);
+
+        this.state = {loginSuccess: false};
+    }
+    onFacebookLoginClick = async() => {
+        const config = {
+            return_scopes: true,
+            scope: 'email'
+        };
+
+        FB.login(response => {
+            const {status} = response;
+
+            this.setState({loginSuccess: status === 'connected'});
+        }, config);
+    };
+
+    onFacebookLogoutClick = async() => {
+        FB.logout(response => {
+            console.log(response);
+        });
+    };
+
+    onGoogleLoginClick = () => {
+        const auth2 = gapi.auth2.getAuthInstance();
+
+        auth2.signIn();
+    };
+
     render() {
         const {classes} = this.props;
+        const {loginSuccess} = this.state;
+
+        if (loginSuccess) {
+            return <Redirect to="/" push={true} />;
+        }
 
         return (
             <div>
                 <CssBaseline />
                 <AppBar title="Login to Feelsbox Manager" includeNav={false} />
                 <div className={classes.root}>
-                    <div className="g-signin2" data-onsuccess="onSignIn" data-theme="dark"></div>
-                    <CircularProgress className={this.props.classes.progress} size="50%" />
+                    <Typography component="p" gutterBottom={true} paragraph={true}>
+                        To begin using your Feelsbox, login using either Facebook or Google.
+                    </Typography>
+                    <Button variant="contained" onClick={this.onFacebookLoginClick} startIcon={<Icon className="fab fa-facebook" />}>Login with Facebook</Button>
+                    <Button variant="contained" onClick={this.onFacebookLogoutClick} startIcon={<Icon className="fab fa-facebook" />}>Logout of Facebook</Button>
+                    <Button variant="contained" onClick={this.onGoogleLoginClick} startIcon={<Icon className="fab fa-google" />}>Login with Google</Button>
                 </div>
             </div>
         );
