@@ -1,26 +1,14 @@
 const path = require('path');
-const HtmlWebPackPlugin = require("html-webpack-plugin");
-const CopyWebpackPlugin = require('copy-webpack-plugin');
+const fs = require('fs');
+const webpack = require('webpack');
+const HtmlWebPackPlugin = require('html-webpack-plugin');
 const WebpackPwaManifest = require('webpack-pwa-manifest');
 
 const htmlPlugin = new HtmlWebPackPlugin({
-    template: "./src/index.html",
-    filename: "./index.html"
+    template: './src/index.html',
+    filename: './index.html'
 });
 
-/*const copyPlugin = new CopyWebpackPlugin([{
-    from: './favicon.ico',
-    to: './favicon.ico',
-    toType: 'file'
-}, {
-    from: './feelsbox.png',
-    to: './feelsbox.png',
-    toType: 'file'
-}, {
-    from: './manifest.json',
-    to: './manifest.json',
-    toType: 'file'
-}]);*/
 const manifestPlugin = new WebpackPwaManifest({
     name: 'FeelsBox',
     short_name: 'FeelsBox',
@@ -34,20 +22,55 @@ const manifestPlugin = new WebpackPwaManifest({
     }]
 });
 
+const hotModule = new webpack.HotModuleReplacementPlugin();
+
 module.exports = {
     module: {
         rules: [{
             test: /\.(js|jsx)$/,
             exclude: /node_modules/,
             use: {
-                loader: "babel-loader"
+                loader: 'babel-loader'
             }
         }]
     },
-    plugins: [htmlPlugin, manifestPlugin],
+    devtool: 'cheap-module-eval-source-map',
+    devServer: {
+        compress: true,
+        contentBase: 'src',
+        historyApiFallback: true,
+        host: 'feelsbox.local',
+        hot: true,
+        https: {
+            ca: fs.readFileSync(('./cert/ca.pem')),
+            cert: fs.readFileSync('./cert/localhost.pem'),
+            key: fs.readFileSync('./cert/localhost-key.pem')
+        },
+        index: 'App',
+        port: 8081,
+        proxy: {
+            '/api/**': {
+                target: 'https://localhost:3000',
+                secure: false
+            }
+        }
+    },
+    output: {
+        pathinfo: false,
+        path: path.resolve(process.cwd(), 'build'),
+        publicPath: '/',
+        filename: '[name].js'
+    },
+    optimization: {
+        splitChunks: {
+            chunks: 'all'
+        }
+    },
+    plugins: [htmlPlugin, manifestPlugin, hotModule],
     resolve: {
         alias: {
-            '-': path.join(__dirname, 'src')
+            '-': path.join(__dirname, 'src'),
+            'react-dom': '@hot-loader/react-dom'
         },
         extensions: ['.js', '.jsx']
     }
